refactor: revert agent-policy back to agent-configuration
The agent-policy rename changed the package's meaning (not just
de-vendoring) and wasn't well justified, so revert it. The other five
renames stand.
