+++ packages/agent-configuration/src/git/claude.ts
