+++ packages/agent-configuration/src/git/opencode.ts
