+++ packages/agent-configuration/src/mcp/index.ts
