+++ packages/agent-configuration/tsconfig.json
