+++ packages/web/lib/mcp/agent-servers.ts
@@ -20,7 +20,7 @@ import {
   GITHUB_MCP_QUALIFIED_NAME,
   safeServerName,
 } from "@background-agents/mcp"
-import type { AgentMcpServer } from "@background-agents/agent-policy/mcp"
+import type { AgentMcpServer } from "@background-agents/agent-configuration/mcp"
 import { type McpOwner, ownerWhere } from "./owner"
 
 // Lazily-initialized GitHub provider
