+++ packages/web/next.config.ts
@@ -7,7 +7,7 @@ const nextConfig: NextConfig = {
   // the string constants, which webpack can tree-shake when not transpiled.
   transpilePackages: [
     "@background-agents/sdk",
-    "@background-agents/agent-policy",
+    "@background-agents/agent-configuration",
     "@background-agents/common",
   ],
   // Keep @daytonaio/sdk on the server side
